+++ lib/mailchimp_api/collection_parsers/segment.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module MailchimpAPI::CollectionParsers
+  class Segment < Base
+  end
+end