+++ CHANGELOG.md
@@ -9,4 +9,5 @@
 - Support of 'lists/<id>/interest-categories/<id>/interests'
 - Support of '/lists/<id>/members'
 - Support of '/lists/<id>/members/<id>/notes'
+- Support of '/lists/<id>/segments'
 - `.count` support